# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Larynx word etymology and language API.",
        "title": "Larynx API",
        "contact": {},
        "version": "preview"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/words": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns English words whose term starts with the given prefix.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "words"
                ],
                "summary": "Search for words",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Prefix to search for",
                        "name": "prefix",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/words/{word}/etymology": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the graph of ancestor words, their language families, and a GeoJSON map of where those languages are spoken.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "words"
                ],
                "summary": "Get a word's etymology graph",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The word to look up",
                        "name": "word",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "English",
                        "description": "Language of the word",
                        "name": "lang",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "true",
                        "description": "Include geojson in the response",
                        "name": "geojson",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "true",
                        "description": "Include the language family tree in the response",
                        "name": "family",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.etymologyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/words/{word}/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a one-sentence summary of a word's origin and history.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "words"
                ],
                "summary": "Get a word's origin and history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The word to look up",
                        "name": "word",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "English",
                        "description": "Language of the word",
                        "name": "lang",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.historyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "server.etymologyResponse": {
            "type": "object",
            "properties": {
                "familyTree": {
                    "$ref": "#/definitions/server.familyNode"
                },
                "geojson": {
                    "$ref": "#/definitions/server.geoJSON"
                },
                "graph": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {}
                    }
                }
            }
        },
        "server.familyNode": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.familyNode"
                    }
                },
                "glottocode": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "server.feature": {
            "type": "object",
            "properties": {
                "geometry": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "properties": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "type": {
                    "type": "string",
                    "example": "Feature"
                }
            }
        },
        "server.geoJSON": {
            "type": "object",
            "properties": {
                "features": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.feature"
                    }
                },
                "type": {
                    "type": "string",
                    "example": "FeatureCollection"
                }
            }
        },
...
        "server.historyResponse": {
            "type": "object",
            "properties": {
                "history": {
                    "type": "string"
                },
                "word": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminJWTAuth": {
            "description": "Admin JWT for blog write endpoints (create/update/delete)",
            "type": "apiKey",
            "name": "X-Admin-JWT",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Client bearer token. Enter the full value as: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}